Move deprecation functions to one spot
